requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Data Handling
pandas>=2.0.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Metadata Extraction
piexif>=1.1.3
exifread>=3.0.0

# Utilities
pyyaml>=6.0
